+++ backend/canisters/user/impl/src/model/direct_chat.rs
@@ -129,7 +129,7 @@ impl DirectChat {
             archived: self.archived.value,
             events_ttl: events_ttl.value,
             events_ttl_last_updated: events_ttl.timestamp,
-            video_call_in_progress: self.events.video_call_in_progress().value.clone(),
+            video_call_in_progress: self.events.video_call_in_progress(Some(my_user_id)),
         }
     }
 
@@ -171,12 +171,7 @@ impl DirectChat {
                 .copied()
                 .map_or(OptionUpdate::NoChange, OptionUpdate::from_update),
             events_ttl_last_updated: (events_ttl.timestamp > updates_since).then_some(events_ttl.timestamp),
-            video_call_in_progress: self
-                .events
-                .video_call_in_progress()
-                .if_set_after(updates_since)
-                .cloned()
-                .map_or(OptionUpdate::NoChange, OptionUpdate::from_update),
+            video_call_in_progress: self.events.video_call_in_progress_updates(Some(my_user_id), updates_since),
         }
     }