+++ backend/canisters/user/impl/src/model/streak.rs
@@ -14,7 +14,7 @@ pub struct Streak {
     insurance_last_updated: TimestampMillis,
     days_insured: u8,
     days_missed: u8,
-    #[serde(skip_deserializing)]
+    #[serde(skip)]
     payment_lock: bool,
     payments: Vec<UserCanisterStreakInsurancePayment>,
     claims: Vec<UserCanisterStreakInsuranceClaim>,